requirements.txt
absl-py==1.2.0
aiofiles==23.2.1
alembic==1.9.2
annotated-types==0.7.0
anyio==4.4.0
arabic-reshaper==3.0.0
asttokens==2.4.1
astunparse==1.6.3
beautifulsoup4==4.12.3
bitarray==2.9.2
black==24.10.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.5
contourpy==1.0.5
cryptography==43.0.1
cycler==0.11.0
Cython==0.29.32
databricks-cli==0.17.4
decorator==5.1.1
diacritization-evaluation==0.5
distro==1.9.0
docker==6.0.1
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.1.0
ez-setup==0.9
farasapy==0.0.14
fastapi==0.112.1
ffmpy==0.4.0
filelock==3.15.4
Flask==2.2.2
flatbuffers==2.0.7
fonttools==4.37.2
fsspec==2024.6.1
gast==0.4.0
gdown==5.2.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gradio==4.41.0
gradio_client==1.3.0
graphviz==0.20.1
greenlet==2.0.1
grpcio==1.48.1
h11==0.14.0
h5py==3.7.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
ibm-cos-sdk==2.13.6
ibm-cos-sdk-core==2.13.6
ibm-cos-sdk-s3transfer==2.13.6
ibm_watsonx_ai==1.1.14
idna==3.3
importlib-metadata==5.2.0
importlib_resources==6.4.3
ipython==8.28.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jiter==0.5.0
jmespath==1.0.1
joblib==1.2.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
llvmlite==0.39.1
lomond==0.3.3
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mlflow==2.1.1
mpmath==1.3.0
mypy-extensions==1.0.0
-e git+https://github.com/NVIDIA-AI-IOT/nanoowl@cfef75a8ad5fb8be0e3beb501a763661a9336d1d#egg=nanoowl
networkx==3.4
numba==0.56.4
numpy==1.23.3
oauthlib==3.2.0
openai==1.41.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
orjson==3.10.7
packaging==24.1
pandas==1.4.4
parso==0.8.4
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.4
pdfreader==0.1.15
Pillow==9.2.0
platformdirs==4.3.6
plotly==5.10.0
plum-dispatch==1.7.4
prompt_toolkit==3.0.48
protobuf==3.19.4
pure_eval==0.2.3
PyArabic==0.6.15
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.6.0
pyparsing==3.0.9
PyPDF2==3.0.1
pypdfium2==4.30.0
PySocks==1.7.1
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2022.2.1
PyWavelets==1.4.1
pywin32==305
PyYAML==6.0
querystring-parser==1.2.4
regex==2024.9.11
requests==2.32.2
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
ruff==0.6.1
safetensors==0.4.5
scikit-learn==1.1.2
scipy==1.9.1
semantic-version==2.10.0
sentence-transformers==3.2.0
sentencepiece==0.2.0
shap==0.41.0
shellingham==1.5.4
six==1.16.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.1
soupsieve==2.6
Spire.Pdf==10.8.1
SQLAlchemy==1.4.46
sqlparse==0.4.3
stack-data==0.6.3
starlette==0.38.2
sympy==1.13.3
tabulate==0.9.0
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tkseem==0.0.3
tokenizers==0.20.1
tomli==2.0.2
tomlkit==0.12.0
torch==2.4.1
tqdm==4.64.1
traitlets==5.14.3
transformers==4.45.2
typer==0.12.4
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
waitress==2.1.2
wcwidth==0.2.13
websocket-client==1.4.2
websockets==12.0
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
